major architectural change
The structure of the program is now a single thread with a select() call to
watch for activity on two file descriptors at once.  Also added the ability
to listen on a Unix domain socket rather than a TCP socket, although this
feature has not yet been tested because the tio-agent needs to have
corresponding changes made to it first.